Fix Bug DeleteUser

+++ CodeStar.Application/Interfaces/IUserServices.cs
@@ -12,7 +12,7 @@ namespace CodeStar.Application.Interfaces
     public interface IUserServices
     {
         public Task<Result<bool>> UserInsert(UserInsertDTO dto);
-        public Task<Result<bool>> DeleteUser(int Id);
+        public Task<Result<bool>> DeleteUser(long Id);
         public Task<Result<UserUpdateDTO>> UpdateUser(int Id,UserUpdateDTO Dto);
         //public Task<Result<GetUserDetailDTO>> GetUserDetail(int Id );
         public Task<Result<bool>> ConfirmEmailAsync(string email, string token);
